Refactor generation of cells to their own files
